grpc-minimal-demo/server/pyproject.toml
[project]
name = "grpc-minimal-demo"
version = "0.1.0"
description = "Minimal gRPC demo — just list devotees from PostgreSQL"
requires-python = ">=3.12"
dependencies = [
    "grpcio>=1.68.0",
    "grpcio-tools>=1.68.0",
    "grpcio-reflection>=1.68.0",
    "protobuf>=5.29.0",
    "fastapi==0.135.1",
    "uvicorn==0.41.0",
    "psycopg2-binary>=2.9.9",
]
